fix polls / interactions timeline loading
